refactor(svelte): split legend/focus pure helpers by domain
Extract entry-key-index, focus-emphasis, and focus-plans from focus.ts so
each file has one pure responsibility. Characterization tests unchanged;
import paths only.
